requirements.txt
numpy
pandas
pyarrow
psutil
scipy
seaborn
matplotlib
jupyter
pytorch==1.8.0
torchvision==0.9.0
torchaudio==0.8.0
cudatoolkit==10.1.243
pytorch-geometric
captum
nilearn